Updating as per suggestions, moving things into an 'internals' module which is optionally made public
