--- src/danger.rs
@@ -1,6 +1,0 @@
-pub use crate::key::blind;
-pub use crate::key::check_public;
-pub use crate::key::decrypt;
-pub use crate::key::decrypt_and_check;
-pub use crate::key::encrypt;
-pub use crate::key::unblind;
